scripts/requirements.txt
psycopg2-binary==2.9.9
pystac-client==0.7.6
geopandas==0.14.2
rasterstats==0.19.0
rasterio==1.3.9
python-dotenv==1.0.0
requests==2.31.0
shapely==2.0.2
pyproj==3.6.1
numpy==1.26.3
pandas==2.1.4
sqlalchemy==2.0.25
geoalchemy2==0.14.3
